Feed new song into sink on current song end
This commit is the base setup for the sink automatically being
fed songs from the queue on songs ending. This is quite difficult
because the song finish function is handled as a callback. When
a source ends, the callback is run. However, when we add a new song
into the sink from the queue, we need to put the same callback on
the end of that song. But, closures do not let us work recursively.

To circumvent this, the closure simply signals an event channel
that a song has ended. Right now, this channel simply is for songs
ending, but in future might be for other events too. A separate thread
waits on this signal, and when the signal is given, this thread is
what adds a new song into the sink. This function can then
just clone the event sender, avoiding the need to make recursive
references to the callback function