Remove volume change event being emitted
This removes a cylic event firing, where the frontned signals a volume
change to the backend, which would then cause the volume to change and
signal the frontend. Instead, this now assumes that the frontend will
know that the volume has changed when it signals it, and will
consequently handle it itself.