Prevent double event on skip
Skipping a song was causing a queue change event to get fired twice.
Problematically, this cause out-of-sync behaviour, because there
is a race condition in how I have coded the skip function and the
queue itself. This fixes the issue caused when skipping by removing
the double emission, but the underlying race condition is still there